requirements.txt
facenet-pytorch
torch
torchvision
tensorflow
opencv-python
uvicorn
fastapi
pillow
numpy
python-multipart
